move infinit functions to misc/infinit_helper
